+++ lib/src/extensions/timestamp_ext.dart
@@ -1,0 +1,7 @@
+import 'package:bson/bson.dart';
+
+extension TimestampExt on Timestamp {
+  bool isAfter(Timestamp other) =>
+      seconds > other.seconds ||
+      (seconds == other.seconds && increment > other.increment);
+}