+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Using Union Base Classes and Error Base classes from type_utils packahge
 - Direct Connection
 - Server Monitoring (no log yet)
+- Server Description
 
 ## 0.1.0-1.1.dev